refactor: global files and directories renaming
