+++ src/app/App.tsx
@@ -1,14 +1,14 @@
 import { useEffect } from 'react';
 import { BrowserRouter, Route, Routes } from 'react-router-dom';
 
-import { Error404Page } from '../pages/Error';
-import { FavoritesPage } from '../pages/Favorites';
-import { LoginPage } from '../pages/Login';
-import { MainPage } from '../pages/Main';
-import { OfferPage } from '../pages/Offer';
-import { Spinner } from '../shared/ui/Spinner';
+import { Error404Page } from '../pages/error';
+import { FavoritesPage } from '../pages/favorites';
+import { LoginPage } from '../pages/login';
+import { MainPage } from '../pages/main';
+import { OfferPage } from '../pages/offer';
+import { Spinner } from '../shared/ui/spinner';
 import { AuthorizationStatus } from './consts';
-import { LoggedRoute } from './LoggedRoute';
+import { LoggedRoute } from './logged-route';
 import { AppRoutes } from './routes';
 import { checkLogin } from './store/api-actions';
 import { useAppDispatch, useAppSelector } from './store/hooks';
