+++ src/app/api/axios-config.ts
@@ -1,2 +1,2 @@
-export const baseUrl = 'https://14.design.htmlacademy.pro/six-cities';
-export const timeout = 5000;
+export const BASE_URL = 'https://14.design.htmlacademy.pro/six-cities';
+export const TIMEOUT = 5000;
