+++ src/app/index.tsx
@@ -1,1 +1,1 @@
-export { App } from './App';
+export { App } from './app';
