+++ src/app/store/interfaces.ts
@@ -1,15 +1,14 @@
-import { City } from '../../entities/City';
-import { OfferCardEntity } from '../../entities/OfferCard';
-import { OfferMaximum } from '../../entities/OfferCard/interfaces';
-import { CommentGet } from '../../entities/Review/interfaces';
-import { User } from '../../entities/User/interfaces';
+import { City } from '../../entities/city';
+import { OfferMaximum, OfferPreview } from '../../entities/offer-card';
+import { CommentGet } from '../../entities/review/interfaces';
+import { User } from '../../entities/user/interfaces';
 import { AuthorizationStatus, FetchStatus } from '../consts';
 import { store } from '.';
 
 export type State = {
   city: City;
 
-  offers?: OfferCardEntity[];
+  offers?: OfferPreview[];
   offersFetchStatus: FetchStatus;
 
   user?: User;
@@ -21,7 +20,7 @@ export type State = {
   reviews?: CommentGet[];
   reviewsFetchStatus: FetchStatus;
 
-  activeOfferId?: OfferCardEntity['id'];
+  activeOfferId?: OfferPreview['id'];
 };
 
 export type AppDispatch = typeof store.dispatch;
