+++ src/entities/City/index.ts
@@ -1,2 +1,4 @@
 export type { CityName, City } from './interfaces';
+export { CityNames } from './interfaces';
 export { cities } from './const';
+export { CitiesList } from './cities-list';
