+++ src/entities/Header/index.ts
@@ -1,1 +1,1 @@
-export { Header } from './ui';
+export { Header } from './header';
