+++ src/entities/Map/ui/index.tsx
@@ -2,10 +2,10 @@ import { layerGroup, Marker } from 'leaflet';
 import { useEffect, useRef } from 'react';
 
 import { useAppSelector } from '../../../app/store/hooks';
-import { City } from '../../City';
+import { City } from '../../city';
 import { Point } from '../interfaces';
-import { useMap } from '../model/useMap';
-import { currentCustomIcon, defaultCustomIcon } from './MapIcons';
+import { useMap } from '../model/use-map';
+import { currentCustomIcon, defaultCustomIcon } from './map-icons';
 
 type MapProps = {
   city: City;
