+++ src/entities/Map/ui/map-icons.ts
