--- src/entities/OfferCard/index.tsx
@@ -1,4 +1,0 @@
-export type { OfferPreview as OfferCardEntity } from './interfaces';
-export { OfferCard } from './ui/OfferCard';
-export { OffersByCities } from './ui/OffersByCities';
-export { OffersList } from './ui/OffersList';
