+++ src/entities/offer-card/index.tsx
@@ -1,0 +1,4 @@
+export type { OfferPreview, OfferMaximum } from './interfaces';
+export { OfferCard } from './ui/offer-card';
+export { OffersByCities } from './ui/offers-by-cities';
+export { OffersList } from './ui/offers-list';
