+++ src/entities/offer-card/interfaces.ts
@@ -1,6 +1,6 @@
 import { components } from '../../../types/schema';
 import { MakeAllRequired } from '../../shared/interfaces';
-import { CityName } from '../City';
+import { CityName } from '../city';
 
 export type OfferType = 'Apartment' | 'Room';
