+++ src/entities/offer-card/ui/other-places-nearby.tsx
