+++ src/index.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import ReactDOM from 'react-dom/client';
 import { Provider } from 'react-redux';
 
-import { App } from './app/App';
+import { App } from './app/app';
 import { store } from './app/store';
 
 const root = ReactDOM.createRoot(
